+++ lib/src/containers/list_tile.dart
@@ -1,14 +1,13 @@
 import 'package:arna/arna.dart';
 
+/// A single fixed-height row that typically contains some text as well as
+/// a leading or trailing icon.
+/// List tiles are typically used in [ListView]s, or arranged in [Column]s.
+/// See also:
+///  * [ArnaCheckBoxListTile], [ArnaRadioListTile], and [ArnaSwitchListTile], widgets
+///    that combine [ListTile] with other controls.
 class ArnaListTile extends StatefulWidget {
-  final Widget? leading;
-  final String? title;
-  final String? subtitle;
-  final Widget? trailing;
-  final GestureTapCallback? onTap;
-  final MouseCursor cursor;
-  final String? semanticLabel;
-
+  /// Creates a list tile.
   const ArnaListTile({
     Key? key,
     this.leading,
@@ -16,9 +15,35 @@ class ArnaListTile extends StatefulWidget {
     this.subtitle,
     this.trailing,
     this.onTap,
+    this.actionable = false,
     this.cursor = MouseCursor.defer,
     this.semanticLabel,
   }) : super(key: key);
+
+  /// A widget to display before the title.
+  final Widget? leading;
+
+  /// The primary content of the list tile.
+  final String? title;
+
+  /// Additional content displayed below the title.
+  final String? subtitle;
+
+  /// A widget to display after the title.
+  final Widget? trailing;
+
+  /// Called when the user taps this list tile.
+  final GestureTapCallback? onTap;
+
+  /// Whether to show disable state or not.
+  final bool actionable;
+
+  /// The cursor for a mouse pointer when it enters or is hovering over the
+  /// list tile.
+  final MouseCursor cursor;
+
+  /// The semantic label of the list tile.
+  final String? semanticLabel;
 
   @override
   State<ArnaListTile> createState() => _ArnaListTileState();
@@ -54,7 +79,26 @@ class _ArnaListTileState extends State<ArnaListTile> {
             if (widget.title != null)
               Padding(
                 padding: Styles.tileTextPadding,
-                child: Row(children: [Flexible(child: Text(widget.title!))]),
+                child: Row(
+                  children: [
+                    Flexible(
+                      child: Text(
+                        widget.title!,
+                        style: ArnaTheme.of(context)
+                            .textTheme
+                            .buttonTextStyle
+                            .copyWith(
+                              color: ArnaDynamicColor.resolve(
+                                !isEnabled && widget.actionable
+                                    ? ArnaColors.disabledColor
+                                    : ArnaColors.primaryTextColor,
+                                context,
+                              ),
+                            ),
+                      ),
+                    ),
+                  ],
+                ),
               ),
             if (widget.subtitle != null)
               Padding(
@@ -64,8 +108,17 @@ class _ArnaListTileState extends State<ArnaListTile> {
                     Flexible(
                       child: Text(
                         widget.subtitle!,
-                        style:
-                            ArnaTheme.of(context).textTheme.subtitleTextStyle,
+                        style: ArnaTheme.of(context)
+                            .textTheme
+                            .subtitleTextStyle
+                            .copyWith(
+                              color: ArnaDynamicColor.resolve(
+                                !isEnabled && widget.actionable
+                                    ? ArnaColors.disabledColor
+                                    : ArnaColors.secondaryTextColor,
+                                context,
+                              ),
+                            ),
                       ),
                     ),
                   ],
@@ -100,7 +153,9 @@ class _ArnaListTileState extends State<ArnaListTile> {
               curve: Styles.basicCurve,
               color: ArnaDynamicColor.resolve(
                 !isEnabled
-                    ? ArnaColors.cardColor
+                    ? widget.actionable
+                        ? ArnaColors.backgroundDisabledColor
+                        : ArnaColors.cardColor
                     : _hover
                         ? ArnaColors.cardHoverColor
                         : ArnaColors.cardColor,