+++ lib/src/inputs/checkbox_list_tile.dart
@@ -1,6 +1,6 @@
 import 'package:arna/arna.dart';
 
-class ArnaCheckBoxListTile extends StatefulWidget {
+class ArnaCheckBoxListTile extends StatelessWidget {
   final bool? value;
   final ValueChanged<bool?>? onChanged;
   final String? title;
@@ -28,135 +28,44 @@ class ArnaCheckBoxListTile extends StatefulWidget {
     this.semanticLabel,
   }) : super(key: key);
 
-  @override
-  State<ArnaCheckBoxListTile> createState() => _ArnaCheckBoxListTileState();
-}
-
-class _ArnaCheckBoxListTileState extends State<ArnaCheckBoxListTile> {
-  bool _hover = false;
-
-  bool get isEnabled => widget.onChanged != null;
-
   void _handleTap() {
-    if (isEnabled) {
-      switch (widget.value) {
+    if (onChanged != null) {
+      switch (value) {
         case false:
-          widget.onChanged!(true);
+          onChanged!(true);
           break;
         case true:
-          widget.onChanged!(widget.tristate ? null : false);
+          onChanged!(tristate ? null : false);
           break;
         case null:
-          widget.onChanged!(false);
+          onChanged!(false);
           break;
       }
     }
   }
 
-  void _handleEnter(event) {
-    if (mounted) setState(() => _hover = true);
-  }
-
-  void _handleExit(event) {
-    if (mounted) setState(() => _hover = false);
-  }
-
   @override
   Widget build(BuildContext context) {
-    return MouseRegion(
-      cursor: widget.cursor,
-      onEnter: _handleEnter,
-      onExit: _handleExit,
-      child: GestureDetector(
-        behavior: HitTestBehavior.opaque,
-        onTap: _handleTap,
-        child: AnimatedContainer(
-          duration: Styles.basicDuration,
-          curve: Styles.basicCurve,
-          color: ArnaDynamicColor.resolve(
-            !isEnabled
-                ? ArnaColors.backgroundDisabledColor
-                : _hover
-                    ? ArnaColors.cardHoverColor
-                    : ArnaColors.cardColor,
-            context,
-          ),
-          padding: Styles.tilePadding,
-          child: Row(
-            children: [
-              Padding(
-                padding: Styles.small,
-                child: ArnaCheckBox(
-                  value: widget.value,
-                  tristate: widget.tristate,
-                  onChanged: widget.onChanged,
-                  isFocusable: widget.isFocusable,
-                  autofocus: widget.autofocus,
-                  accentColor: widget.accentColor,
-                  cursor: widget.cursor,
-                  semanticLabel: widget.semanticLabel,
-                ),
-              ),
-              Expanded(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    if (widget.title != null)
-                      Padding(
-                        padding: Styles.tileTextPadding,
-                        child: Row(
-                          children: [
-                            Flexible(
-                              child: Text(
-                                widget.title!,
-                                style: ArnaTheme.of(context)
-                                    .textTheme
-                                    .buttonTextStyle
-                                    .copyWith(
-                                      color: ArnaDynamicColor.resolve(
-                                        !isEnabled
-                                            ? ArnaColors.disabledColor
-                                            : ArnaColors.primaryTextColor,
-                                        context,
-                                      ),
-                                    ),
-                              ),
-                            ),
-                          ],
-                        ),
-                      ),
-                    if (widget.subtitle != null)
-                      Padding(
-                        padding: Styles.tileTextPadding,
-                        child: Row(
-                          children: [
-                            Flexible(
-                              child: Text(
-                                widget.subtitle!,
-                                style: ArnaTheme.of(context)
-                                    .textTheme
-                                    .buttonTextStyle
-                                    .copyWith(
-                                      color: ArnaDynamicColor.resolve(
-                                        !isEnabled
-                                            ? ArnaColors.disabledColor
-                                            : ArnaColors.secondaryTextColor,
-                                        context,
-                                      ),
-                                    ),
-                              ),
-                            ),
-                          ],
-                        ),
-                      ),
-                  ],
-                ),
-              ),
-              if (widget.trailingButton != null) widget.trailingButton!,
-            ],
-          ),
+    return ArnaListTile(
+      leading: Padding(
+        padding: Styles.small,
+        child: ArnaCheckBox(
+          value: value,
+          tristate: tristate,
+          onChanged: onChanged,
+          isFocusable: isFocusable,
+          autofocus: autofocus,
+          accentColor: accentColor,
+          cursor: cursor,
+          semanticLabel: semanticLabel,
         ),
       ),
+      title: title,
+      subtitle: subtitle,
+      trailing: trailingButton,
+      onTap: onChanged != null ? _handleTap : null,
+      actionable: true,
+      cursor: cursor,
     );
   }
 }